Bounds calculations for gutters in grouping box
If there is a layer group, include its bounds as 
well so that its internal layers gets missed out 
at that level

If there is a layer group but the grouping box has
exactly same layers as the layer group, then that 
is the render layer group. Ignore group details 
and group the layers as usual.